Python/requirements.txt
certifi
chardet
cssselect
icalendar
idna
lxml
Pillow
pytesseract
python-dateutil
pytz
requests
six
urllib3
